+++ docs/custom.tsx
@@ -1,5 +1,0 @@
-export const pageType = 'custom';
-
-export default function Custom() {
-  return <div>This is a custom component </div>;
-}