+++ src/runtime/app.tsx
@@ -3,11 +3,12 @@ import { routes } from 'virtual:routes';
 import { matchRoutes, useLocation } from 'react-router-dom';
 import siteData from 'island:site-data';
 import { Route } from '../node/plugin-routes';
-import { cleanUrl, getRelativePagePath, omit } from './utils';
+import { getRelativePagePath, omit } from './utils';
 import { PageData } from '../shared/types';
 import { HelmetProvider } from 'react-helmet-async';
 import { useContext, useLayoutEffect } from 'react';
 import { DataContext } from 'island/client';
+import { cleanUrl } from '../shared/utils';
 
 export async function waitForApp(path: string): Promise<PageData> {
   const matched = matchRoutes(routes, path, siteData.base)!;
@@ -18,8 +19,9 @@ export async function waitForApp(path: string): Promise<PageData> {
 
     const pagePath = cleanUrl((matched[0].route as Route).filePath);
     const relativePagePath = getRelativePagePath(path, pagePath, siteData.base);
+    const isApiPage = mod?.meta?.api || mod?.meta?.pageType === 'api';
     // API Page
-    if (mod?.meta?.api || mod?.meta?.pageType === 'api') {
+    if (isApiPage) {
       const subModules = await Promise.all(
         routes
           .filter(