+++ src/runtime/utils.ts
@@ -1,4 +1,4 @@
-export const inBrowser = () => typeof window !== 'undefined';
+import { cleanUrl } from '../shared/utils';
 
 export const omit = (obj: Record<string, unknown>, keys: string[]) => {
   const ret = { ...obj };
@@ -7,12 +7,6 @@ export const omit = (obj: Record<string, unknown>, keys: string[]) => {
   }
   return ret;
 };
-
-export const queryRE = /\?.*$/s;
-export const hashRE = /#.*$/s;
-
-export const cleanUrl = (url: string): string =>
-  url.replace(hashRE, '').replace(queryRE, '');
 
 export const getRelativePagePath = (
   routePath: string,