+++ src/shared/utils/index.ts
@@ -1,0 +1,7 @@
+export const queryRE = /\?.*$/s;
+export const hashRE = /#.*$/s;
+
+export const cleanUrl = (url: string): string =>
+  url.replace(hashRE, '').replace(queryRE, '');
+
+export const inBrowser = () => typeof window !== 'undefined';