+++ src/theme-default/components/DocFooter/index.tsx
@@ -1,13 +1,25 @@
 import styles from './index.module.scss';
 import { usePageData } from 'island/client';
-import { useEditLink, usePrevNextPage } from '../../logic';
+import { useEditLink, useLocaleSiteData, usePrevNextPage } from '../../logic';
 import { normalizeHref } from '../../logic/index';
+import { useLocation } from 'react-router-dom';
 
 export function DocFooter() {
-  const { siteData, relativePagePath } = usePageData();
+  const { siteData, relativePagePath, lastUpdatedTime } = usePageData();
   const { prevPage, nextPage } = usePrevNextPage(siteData);
-  const { editLink: rawEditLink } = siteData.themeConfig;
-  const editLink = useEditLink(rawEditLink!, relativePagePath);
+  const { pathname } = useLocation();
+  const themeConfig = siteData.themeConfig;
+  const {
+    editLink: rawEditLink,
+    lastUpdatedText,
+    prevPageText = 'Previous Page',
+    nextPageText = 'Next page'
+  } = useLocaleSiteData(themeConfig, pathname);
+
+  const editLink = useEditLink(
+    rawEditLink! ?? themeConfig?.editLink,
+    relativePagePath
+  );
 
   return (
     <footer className={styles.footer}>
@@ -20,17 +32,16 @@ export function DocFooter() {
           </div>
         ) : null}
 
-        {/* TODO */}
-        {/* <div className={styles.lastUpdated}>
-          {lastUpdatedText ? (
+        <div className={styles.lastUpdated}>
+          {
             <>
               <p className={styles.lastUpdated}>
-                {editLink?.text || 'Last Updated: '}
+                {`${lastUpdatedText ?? 'Last Updated'}: `}
               </p>
-              <span>{}</span>
+              <span>{lastUpdatedTime}</span>
             </>
-          ) : null}
-        </div> */}
+          }
+        </div>
       </div>
 
       <div className={styles.prevNext}>
@@ -40,7 +51,7 @@ export function DocFooter() {
               href={normalizeHref(prevPage.link)}
               className={`${styles.pagerLink} ${styles.prev}`}
             >
-              <span className={styles.desc}>Previous Page</span>
+              <span className={styles.desc}>{prevPageText}</span>
               <span className={styles.title}>{prevPage.text}</span>
             </a>
           ) : null}
@@ -52,7 +63,7 @@ export function DocFooter() {
                 href={normalizeHref(nextPage.link)}
                 className={`${styles.pagerLink} ${styles.next}`}
               >
-                <span className={styles.desc}>Next Page</span>
+                <span className={styles.desc}>{nextPageText}</span>
                 <span className={styles.title}>{nextPage.text}</span>
               </a>
             </div>