+++ src/theme-default/components/SwitchAppearance/index.tsx
@@ -3,13 +3,11 @@ import styles from './index.module.scss';
 import SunSvg from './sun.svg';
 import MoonSvg from './moon.svg';
 import { ComponentPropsWithIsland } from '../../../shared/types';
-import { useAppearance } from '../../logic';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export function SwitchAppearance(_props: ComponentPropsWithIsland) {
-  const handleClick = useAppearance();
   return (
-    <Switch onClick={handleClick}>
+    <Switch id="appearance">
       <SunSvg className={styles.sun}></SunSvg>
       <MoonSvg className={styles.moon}></MoonSvg>
     </Switch>